shrink_resize that not work

+++ src/latentflow/slice_utils.py
@@ -2,12 +2,12 @@ import logging
 def slice_scale(s, q):
     result = None
     if isinstance(s, slice):
-        result = slice(s.start*q, s.stop*q)
+        result = slice(int(s.start*q), int(s.stop*q))
     elif isinstance(s, list):
         l = []
         for x in s:
             for i in range(q):
-                l.append(x*q + i)
+                l.append(int(x*q + i))
         result = l
     else:
         raise TypeError(f"Uknown type {type(s)}")
